fix: issue in query which leads to no locks removed
This commit fixes an issue with the query logic which prevented locks
from being deleted as long as more than one lock is expired.

To prevent a potential mass deletion of data due to the fix, the code
has been adjusted to delete only 1000 locks at every run of the
background job.

Signed-off-by: Salvatore Martire <4652631+salmart-dev@users.noreply.github.com>